Rename command, add more tests
