+++ config/params.php
@@ -16,7 +16,7 @@ use Yiisoft\Yii\Debug\Collector\RouterCollector;
 use Yiisoft\Yii\Debug\Collector\RouterCollectorInterface;
 use Yiisoft\Yii\Debug\Collector\ServiceCollectorInterface;
 use Yiisoft\Yii\Debug\Collector\WebAppInfoCollector;
-use Yiisoft\Yii\Debug\Command\ClearCommand;
+use Yiisoft\Yii\Debug\Command\ResetCommand;
 use Yiisoft\Yii\Debug\Proxy\ContainerProxy;
 use Yiisoft\Yii\Debug\Proxy\EventDispatcherInterfaceProxy;
 use Yiisoft\Yii\Debug\Proxy\LoggerInterfaceProxy;
@@ -58,7 +58,7 @@ return [
     ],
     'yiisoft/yii-console' => [
         'commands' => [
-            'debug/reset' => ClearCommand::class,
+            'debug/reset' => ResetCommand::class,
         ],
     ],
 ];
